requirements.txt
Pillow==10.4.0
tkinterdnd2==0.3.0
customtkinter==5.2.2
CTkToolTip==0.8
plyer==2.1.0
ttkbootstrap==1.10.1
